changes to Performance Use case

+++ src/main/java/edu/cmu/sv/ws/ssnoc/data/dao/IMessageDAO.java
@@ -32,9 +32,9 @@ public interface IMessageDAO {
 
     List<ExchangeInfoPO> loadChatBuddies(String userName);
 
-    int getGetWallRequestsCount();
+    long getGetWallRequestsCount();
 
-    int getPostWallRequestCount();
+    long getPostWallRequestCount();
 
     void resetRequestsCount();
